requirements.txt
certifi==2024.8.30
charset-normalizer==3.3.2
idna==3.8
requests==2.31.0
urllib3==2.2.3
websockets==12.0
twilio==8.12.0
python-dotenv==1.0.0
tinydb==4.8.0
python-dateutil==2.8.2
# langchain-community
# PyMuPDF
trieve_py_client
google-cloud-storage
pydantic>=2.0.0
pymupdf4llm